refactor: `util` 패키지 하위 클래스 `ui.util`로 이동
