+++ android/app/src/main/java/com/mulkkam/ui/history/HistoryFragment.kt
@@ -24,7 +24,7 @@ import com.mulkkam.ui.history.adapter.HistoryAdapter
 import com.mulkkam.ui.history.adapter.HistoryViewHolder
 import com.mulkkam.ui.history.dialog.DeleteConfirmDialogFragment
 import com.mulkkam.ui.main.Refreshable
-import com.mulkkam.ui.util.getColoredSpannable
+import com.mulkkam.ui.util.extensions.getColoredSpannable
 import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
