+++ android/app/src/main/java/com/mulkkam/ui/home/HomeFragment.kt
@@ -19,7 +19,7 @@ import com.mulkkam.ui.custom.ExtendableFloatingMenuItem
 import com.mulkkam.ui.home.dialog.ManualDrinkFragment
 import com.mulkkam.ui.main.Refreshable
 import com.mulkkam.ui.notification.NotificationActivity
-import com.mulkkam.ui.util.getColoredSpannable
+import com.mulkkam.ui.util.extensions.getColoredSpannable
 import java.util.Locale
 
 class HomeFragment :
