+++ android/app/src/main/java/com/mulkkam/ui/onboarding/nickname/NicknameFragment.kt
@@ -13,8 +13,8 @@ import com.mulkkam.R
 import com.mulkkam.databinding.FragmentNicknameBinding
 import com.mulkkam.ui.binding.BindingFragment
 import com.mulkkam.ui.onboarding.OnboardingViewModel
-import com.mulkkam.ui.util.getAppearanceSpannable
-import com.mulkkam.util.extensions.applyImeMargin
+import com.mulkkam.ui.util.extensions.applyImeMargin
+import com.mulkkam.ui.util.extensions.getAppearanceSpannable
 
 class NicknameFragment :
     BindingFragment<FragmentNicknameBinding>(
