+++ android/app/src/main/java/com/mulkkam/ui/settingtargetamount/SettingTargetAmountActivity.kt
@@ -14,9 +14,9 @@ import androidx.core.widget.doAfterTextChanged
 import com.mulkkam.R
 import com.mulkkam.databinding.ActivitySettingTargetAmountBinding
 import com.mulkkam.ui.binding.BindingActivity
-import com.mulkkam.ui.util.getAppearanceSpannable
-import com.mulkkam.ui.util.getColoredSpannable
-import com.mulkkam.util.extensions.applyImeMargin
+import com.mulkkam.ui.util.extensions.applyImeMargin
+import com.mulkkam.ui.util.extensions.getAppearanceSpannable
+import com.mulkkam.ui.util.extensions.getColoredSpannable
 import java.util.Locale
 
 class SettingTargetAmountActivity : BindingActivity<ActivitySettingTargetAmountBinding>(ActivitySettingTargetAmountBinding::inflate) {
