+++ android/app/src/main/java/com/mulkkam/ui/util/glide/SvgModule.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.util.glide
+package com.mulkkam.ui.util.glide
 
 import android.content.Context
 import android.graphics.drawable.PictureDrawable
